requirements.txt
# ============================================================================
# Internship Auto-Applier - Python Dependencies
# ============================================================================
# Install: pip install -r requirements.txt
# ============================================================================

# Core
python-dotenv>=1.0.0
pyyaml>=6.0

# Browser Automation
playwright>=1.40.0
playwright-stealth>=1.0.6

# HTTP Requests
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
aiosqlite>=0.19.0

# AI (Google Gemini)
google-generativeai>=0.8.0

# Async
asyncio-throttle>=1.0.0

# CLI & Utils
click>=8.1.0
rich>=13.0.0
python-dateutil>=2.8.0

# Logging
loguru>=0.7.0

# File handling
python-magic>=0.4.27
PyPDF2>=3.0.0

# CAPTCHA solving
2captcha-python>=1.2.0

# Development
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# ruff>=0.1.0
